+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changed
 
 - Value for unlimited item slots in machine from -1 to 255 to use the same value as in-game.
+- Usage of `ReflectionFactory`to `AutoWireFactory`
 
 ## 2.0.0 - 2019-04-07